Fix partial citation normalization
