+++ src/modules/AieraChat/services/messages.ts
@@ -14,6 +14,7 @@ import {
     ChatSessionWithMessagesQuery,
     ChatSessionWithMessagesQueryVariables,
     ChatSource,
+    Citation as RawCitation,
     ConfirmationChatSourceInput,
     ChatSourceType,
     ConfirmChatMessageSourceConfirmationMutation,
@@ -22,14 +23,17 @@ import {
     CreateChatMessagePromptMutationVariables,
     TextBlock,
 } from '@aiera/client-sdk/types/generated';
-import { BlockType, ContentBlock } from '@aiera/client-sdk/modules/AieraChat/components/Messages/MessageFactory/Block';
+import {
+    BlockType,
+    Citation,
+    ContentBlock,
+} from '@aiera/client-sdk/modules/AieraChat/components/Messages/MessageFactory/Block';
 import {
     CHAT_SESSION_QUERY,
     CONFIRM_SOURCE_CONFIRMATION_MUTATION,
     CREATE_CHAT_MESSAGE_MUTATION,
 } from '@aiera/client-sdk/modules/AieraChat/services/graphql';
 import { Source, useChatStore } from '@aiera/client-sdk/modules/AieraChat/store';
-import { normalizeCitation } from '@aiera/client-sdk/modules/AieraChat/services/utils';
 
 const POLLING_INTERVAL = 5000; // 5 seconds
 const MAX_POLLING_DURATION = 60 * 60 * 1000; // 1 hour in milliseconds
@@ -133,6 +137,25 @@ function mapConfirmedSourcesToInput(sources: Source[]): ConfirmationChatSourceIn
         sourceName: source.title,
         sourceType: source.targetType as ChatSourceType,
     }));
+}
+
+/**
+ * Map raw citations from the server
+ */
+export function normalizeCitation(rawCitation: RawCitation): Citation {
+    const source = rawCitation.source;
+    const sourceParent = source.parent as ChatSource;
+    return {
+        author: rawCitation.author || '',
+        contentId: source.sourceId,
+        date: rawCitation.date as string,
+        marker: rawCitation.marker,
+        meta: rawCitation.meta as object,
+        source: source.name,
+        sourceId: sourceParent ? sourceParent.sourceId : source.sourceId,
+        text: rawCitation.quote,
+        url: rawCitation.url || undefined,
+    };
 }
 
 /**
